+++ Notifisor/Notifisor/CoreData/Day+CoreDataProperties.swift
@@ -43,14 +43,3 @@ extension Day {
 extension Day : Identifiable {
 
 }
-
-extension Day: Comparable {
-    public static func <(lhs: Day, rhs: Day) -> Bool {
-        guard let lhsDate = lhs.date,
-              let rhsDate = rhs.date
-        else {
-            fatalError("dates are invalid")
-        }
-        return lhsDate < rhsDate
-    }
-}
